move resolveCdpElementId
